# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/matches/{matchId}/round/{roundId}":{"put":{"tags":["match-controller"],"summary":"Manage a round update (the play of the player)","description":"Return the round updated based on the player decision.If the round or match do not exist, then return 404 not foundIf the match to not belong to the player id received in the request.body then return 400 bad requestIf the round to not belong to the match then return 400 bad request","operationId":"putRoundMatch","parameters":[{"name":"matchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"roundId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoundPlayDTO"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoundResponseDTO"}}}},"400":{"description":"The round id {roundId} does not belong to match {matchId}","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"The round id {roundId} not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"403":{"description":"Round is end","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players":{"post":{"tags":["player-controller"],"summary":"Create a new player","description":"Return the player created with your id. If a player exist with the username or email, then return 400. Additionally the email must be valid.","operationId":"postPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPlayerRequestDTO"}}},"required":true},"responses":{"400":{"description":"The email format it's not valid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerResponseDTO"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/matches":{"post":{"tags":["match-controller"],"summary":"Create a new match","description":"Return the match created with they id. If a player has a PLAYING match, then return that match. If the player do not exist, then return 404 not found","operationId":"postMatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMatchRequestDTO"}}},"required":true},"responses":{"404":{"description":"The player id {playerId} do not exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponseDTO"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/matches/{id}/round/":{"post":{"tags":["match-controller"],"summary":"Create a new Round in a match","description":"Return the round created in the match if the player do not have unfinished rounds, otherwise return the unfinished round.If the match do not exist, then return 404 not foundIf the match to not belong to the player id received in the request.body. then return 400 bad request","operationId":"postRoundMatch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMatchRoundRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoundResponseDTO"}}}},"400":{"description":"The match id {matchId} does not belong to player {playerId}","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"The match id {matchId} not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"403":{"description":"Insufficient balance","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a player by id","description":"Return a player by they id. If the player doesn't exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerResponseDTO"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}/matches":{"get":{"tags":["player-controller"],"summary":"Get all matches of a player","description":"Return all matches belonging to a player.","operationId":"getMatchesOfPlayer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerResponseDTO"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/matches/{id}":{"get":{"tags":["match-controller"],"summary":"Get a match by id","description":"Return a match by they id. If the match doesn't exist return 404","operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponseDTO"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"RoundPlayDTO":{"required":["decision","player_id"],"type":"object","properties":{"player_id":{"type":"integer","format":"int64"},"decision":{"type":"string","enum":["NEW_CARD","STOP"]}}},"Card":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"value":{"type":"number"},"suit":{"type":"string","enum":["OROS","COPAS","ESPADAS","BASTOS"]}}},"RoundResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"match_id":{"type":"integer","format":"int64"},"player_cards":{"type":"array","items":{"$ref":"#/components/schemas/Card"}},"player_hand_status":{"type":"string","enum":["IN_GAME","STOPPED","EXCEEDED"]},"app_cards":{"type":"array","items":{"$ref":"#/components/schemas/Card"}},"app_hand_status":{"type":"string","enum":["IN_GAME","STOPPED","EXCEEDED"]},"chips_in_play":{"type":"number"},"winner":{"type":"string","enum":["PLAYER","APP"]}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"NewPlayerRequestDTO":{"required":["email","password","user_name"],"type":"object","properties":{"password":{"type":"string"},"email":{"type":"string"},"user_name":{"type":"string"}}},"PlayerResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"avatar":{"type":"string"},"balance":{"type":"number"},"user_name":{"type":"string"}}},"NewMatchRequestDTO":{"required":["playerId"],"type":"object","properties":{"playerId":{"type":"integer","format":"int64"}}},"MatchResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rounds":{"type":"array","items":{"$ref":"#/components/schemas/RoundResponseDTO"}},"player":{"$ref":"#/components/schemas/PlayerResponseDTO"},"matchStatus":{"type":"string","enum":["PLAYING","FINISH"]}}},"NewMatchRoundRequestDTO":{"required":["player_id"],"type":"object","properties":{"player_id":{"type":"integer","format":"int64"}}}}}}